blog/requirement.txt
asgiref==3.8.1
asttokens==3.0.0
decorator==5.2.1
Django==5.1.6
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==3.2.3
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
executing==2.2.0
iniconfig==2.1.0
ipython==9.0.1
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.1.7
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pluggy==1.5.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
PyJWT==2.9.0
pytest==8.3.5
pytest-django==4.10.0
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
wcwidth==0.2.13
